+++ playground/core/main.js
@@ -7,28 +7,12 @@ import Swiper from '../../build/swiper-bundle.esm';
 // eslint-disable-next-line
 window.swiper = new Swiper({
   el: '.swiper-container',
-  initialSlide: 2,
-  spaceBetween: 50,
-  slidesPerView: 2,
-  centeredSlides: true,
   slideToClickedSlide: true,
   grabCursor: true,
-  scrollbar: {
-    el: '.swiper-scrollbar',
-  },
-  mousewheel: {
-    enabled: true,
-  },
-  keyboard: {
-    enabled: true,
-  },
-  pagination: {
-    el: '.swiper-pagination',
-  },
-  navigation: {
-    nextEl: '.swiper-button-next',
-    prevEl: '.swiper-button-prev',
-  },
+  scrollbar: true,
+  createElements: true,
+  pagination: true,
+  navigation: true,
   a11y: {
     containerMessage: 'Example content',
     containerRoleDescriptionMessage: 'carousel',