+++ src/components/zoom/zoom.js
@@ -298,15 +298,27 @@ const Zoom = {
     const { gesture, image } = zoom;
 
     if (!gesture.$slideEl) {
-      if (swiper.params.virtual && swiper.params.virtual.enabled && swiper.virtual) {
-        gesture.$slideEl = swiper.$wrapperEl.children(`.${swiper.params.slideActiveClass}`);
-      } else {
-        gesture.$slideEl = swiper.slides.eq(swiper.activeIndex);
+      if (e && e.target) {
+        gesture.$slideEl = $(e.target).closest(`.${swiper.params.slideClass}`);
       }
+      if (!gesture.$slideEl) {
+        if (swiper.params.virtual && swiper.params.virtual.enabled && swiper.virtual) {
+          gesture.$slideEl = swiper.$wrapperEl.children(`.${swiper.params.slideActiveClass}`);
+        } else {
+          gesture.$slideEl = swiper.slides.eq(swiper.activeIndex);
+        }
+      }
+
       gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas, picture, .swiper-zoom-target');
       gesture.$imageWrapEl = gesture.$imageEl.parent(`.${params.containerClass}`);
     }
-    if (!gesture.$imageEl || gesture.$imageEl.length === 0) return;
+    if (
+      !gesture.$imageEl ||
+      gesture.$imageEl.length === 0 ||
+      !gesture.$imageWrapEl ||
+      gesture.$imageWrapEl.length === 0
+    )
+      return;
 
     gesture.$slideEl.addClass(`${params.zoomedSlideClass}`);
 
@@ -398,7 +410,13 @@ const Zoom = {
       gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas, picture, .swiper-zoom-target');
       gesture.$imageWrapEl = gesture.$imageEl.parent(`.${params.containerClass}`);
     }
-    if (!gesture.$imageEl || gesture.$imageEl.length === 0) return;
+    if (
+      !gesture.$imageEl ||
+      gesture.$imageEl.length === 0 ||
+      !gesture.$imageWrapEl ||
+      gesture.$imageWrapEl.length === 0
+    )
+      return;
 
     zoom.scale = 1;
     zoom.currentScale = 1;