+++ src/types/swiper-class.d.ts
@@ -161,7 +161,22 @@ interface Swiper extends SwiperClass<SwiperEvents> {
    */
   allowTouchMove: boolean;
 
+  /**
+   * !INTERNAL
+   */
   rtlTranslate: boolean;
+
+  /**
+   * Disable Swiper (if it was enabled). When Swiper is disabled, it will hide all navigation elements and won't respond to any events and interactions
+   *
+   */
+  disable(): void;
+
+  /**
+   * Enable Swiper (if it was disabled)
+   *
+   */
+  enable(): void;
 
   /**
    * Set Swiper translate progress (from 0 to 1). Where 0 - its initial position (offset) on first slide, and 1 - its maximum position (offset) on last slide
@@ -410,16 +425,36 @@ interface Swiper extends SwiperClass<SwiperEvents> {
    */
   offAny(handler: (eventName: string, ...args: any[]) => void): void;
 
+  /**
+   * !INTERNAL
+   */
   isHorizontal(): boolean;
 
+  /**
+   * !INTERNAL
+   */
   getBreakpoint(breakpoints: SwiperOptions['breakpoints']): string;
 
+  /**
+   * !INTERNAL
+   */
   setBreakpoint(): void;
 
+  /**
+   * !INTERNAL
+   */
   currentBreakpoint: any;
 
+  /**
+   * !INTERNAL
+   */
   destroyed: boolean;
+
+  /**
+   * !INTERNAL
+   */
   modules: Array<any>; //TODO: add typing
+
   a11y: A11yMethods;
   autoplay: AutoplayMethods;
   controller: ControllerMethods;