+++ src/vue/get-params.js
@@ -38,6 +38,7 @@ function getParams(obj = {}) {
 
   ['navigation', 'pagination', 'scrollbar'].forEach((key) => {
     if (params[key] === true) params[key] = {};
+    if (params[key] === false) delete params[key];
   });
 
   return { params, passedParams, rest };